# The RAND_RANGE() macro is immediately overwriting rnd_idx