@nikic 2017-12-09
Thanks for catching this. I've expanded the patch to drop RAND_RANGE() entirely in https://github.com/php/php-src/commit/e4e26f24285511c5aa2a359ff7e93c1748fdda3b.